Move -exp(w) from python to cuda in WKV kernel

+++ RWKV-v4neo/cuda/wkv_op_bf16.cpp
@@ -2,10 +2,10 @@
 #include "ATen/ATen.h"
 typedef at::BFloat16 bf16;
 
-void cuda_forward(int B, int T, int C, float *w, bf16 *u, bf16 *k, bf16 *v, bf16 *y);
+void cuda_forward(int B, int T, int C, bf16 *w, bf16 *u, bf16 *k, bf16 *v, bf16 *y);
 
 void forward(int64_t B, int64_t T, int64_t C, torch::Tensor &w, torch::Tensor &u, torch::Tensor &k, torch::Tensor &v, torch::Tensor &y) {
-    cuda_forward(B, T, C, w.data_ptr<float>(), u.data_ptr<bf16>(), k.data_ptr<bf16>(), v.data_ptr<bf16>(), y.data_ptr<bf16>());
+    cuda_forward(B, T, C, w.data_ptr<bf16>(), u.data_ptr<bf16>(), k.data_ptr<bf16>(), v.data_ptr<bf16>(), y.data_ptr<bf16>());
 }
 
 PYBIND11_MODULE(TORCH_EXTENSION_NAME, m) {
